requirements.txt
Django>=5.1
gunicorn
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
cloudinary
cloudinary-storage
